// src/types/rolePermissions.ts
export enum PermissionValues {
    ROLE_MANAGER = 'role_manager',
    NO_ACCESS = 'no_access',
    OBJECT_ADMINISTRATOR = 'object_administrator',
    OBJECT_MANAGER = 'object_manager',
    OBJECT_READER = 'object_reader',
    OBJECT_SHARER = 'object_sharer',
}

export enum PermissionLabels {
    ROLE_MANAGER = 'Role Manager',
    NO_ACCESS = 'No Access',
    ADMINISTRATOR = 'Administrator',
    CREATOR = 'Creator',
    READ_ONLY = 'Read-Only',
    SHARER = 'Sharer',
}